SIG-517 Big move in test modules to extract all existing tests for email integrations to it's own test package.
